platform/backend/requirements.txt
# CRMBLR Platform Backend Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Authentication
pyjwt==2.8.0
bcrypt==4.1.1
python-jose[cryptography]==3.3.0

# Data Processing
pandas==2.1.3
numpy==1.25.2
openpyxl==3.1.2

# Search & Analytics
elasticsearch==8.11.0

# API Documentation
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Requests
httpx==0.25.2
requests==2.31.0

# Email
fastapi-mail==1.4.1

# Payment Processing
stripe==7.8.0

# File Handling
python-multipart==0.0.6
aiofiles==23.2.1

# Validation & Utilities
email-validator==2.1.0
python-dotenv==1.0.0

# CORS
fastapi-cors==0.0.6

# Template Rendering
jinja2==3.1.2

# Static Files
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Production
gunicorn==21.2.0